# apis/app-api/game/games-app-api.openapi.json
{
  "openapi": "3.1.2",
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "info": {
    "title": "SDKWork Games App API",
    "version": "0.1.0",
    "description": "SDKWork games platform HTTP contract.",
    "x-sdkwork-owner": "sdkwork-games",
    "x-sdkwork-domain": "game"
  },
  "servers": [
    {
      "url": "/app/v3/api",
      "description": "SDKWork API root"
    }
  ],
  "paths": {
    "/app/v3/api/system/health": {
      "get": {
        "operationId": "games.health.check",
        "tags": [
          "health"
        ],
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-owner": "sdkwork-games",
        "x-sdkwork-domain": "game",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesHealthResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        }
      }
    },
    "/app/v3/api/system/ready": {
      "get": {
        "operationId": "games.ready.check",
        "tags": [
          "health"
        ],
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-owner": "sdkwork-games",
        "x-sdkwork-domain": "game",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesHealthResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        }
      }
    },
    "/app/v3/api/games": {
      "get": {
        "operationId": "games.catalog.list",
        "tags": [
          "games"
        ],
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-owner": "sdkwork-games",
        "x-sdkwork-domain": "game",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/PageSizeQuery"
          },
          {
            "$ref": "#/components/parameters/StatusQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesApiResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        }
      }
    },
    "/app/v3/api/games/{gameId}": {
      "get": {
        "operationId": "games.catalog.retrieve",
        "tags": [
          "games"
        ],
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-owner": "sdkwork-games",
        "x-sdkwork-domain": "game",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/GameIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesApiResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SDKWork-Auth-Token",
        "description": "SDKWork dual-token auth principal (Authorization bearer)."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork dual-token access credential header."
      }
    },
    "parameters": {
      "GameIdPath": {
        "name": "gameId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PageQuery": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSizeQuery": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
      "StatusQuery": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ProblemDetailResponse": {
        "description": "RFC 9457 problem details.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      }
    },
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "GamesApiResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
      "GamesHealthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "service"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          }
        }
      },
      "GameCatalogItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "gameCode",
          "title",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "gameCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "GameCatalogPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameCatalogItem"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1
          }
        }
      }
    }
  }
}